Set sla_triaged_at on apropriate report activities
This properly indicates that a report was triaged (for SLA metric purposes)
when a report was marked as duplicate, spam, informative, etc. This
indicates that the H1 triage team saw the issue and took apropriate action,
which is the timestamp we're looking for from an SLA POV.

(The triaged_at date in the H1 API reflects when the issue's state was changed
to "triaged", which isn't exatly right for our purposes. For example, an issue
that gets kicked back to the reporter for more info is "triaged" as far as
metrics are concerned, but stays in the "new" state in H1.)